--- src/js/translations.ts
@@ -1,269 +1,0 @@
-export default {
-  "en-US": {
-    document_title: "Literature Clock",
-    add_quote: "Add quote",
-    report_error: "Report error",
-    made_by: "Made by",
-    project: "Project",
-    work_mode: "Work",
-    zen_mode: "Zen",
-    screensaver_mode: "Screensaver",
-    title: "Title",
-    author: "Author",
-
-    zen_mode_title: "Remove everything that distracts you",
-    work_mode_title: "Show only safe for work quotes",
-    exit_zen_title: "Exit Zen mode",
-
-    theme: "Theme",
-    colors: "Colors",
-    base: "Base",
-    pink: "Pink",
-    green: "Green",
-    orange: "Orange",
-    purple: "Purple",
-    blue: "Blue",
-    gray: "Gray",
-    color: "Random colors",
-
-    themes: "Themes",
-    retro: "Retro",
-    elegant: "Elegant",
-    festive: "Festive",
-    bohemian: "Bohemian",
-    handwriting: "Handwriting",
-    anaglyph: "Anaglyph",
-    whatsapp: "Whatsapp",
-    terminal: "Terminal",
-    frame: "Frame",
-    subtle: "Subtle",
-
-    system: "System",
-    light: "Light",
-    dark: "Dark",
-
-    language: "Language",
-    "en-US": "English (US)",
-    "es-ES": "Spanish (ES)",
-    "pt-BR": "Portuguese (BR)",
-    "fr-FR": "French",
-    "it-IT": "Italian",
-    random: "Random language",
-
-    themeFont: "Theme font",
-  },
-  "es-ES": {
-    document_title: "Reloj Literario",
-    add_quote: "Agregar cita",
-    report_error: "Reportar error",
-    made_by: "Hecho por",
-    project: "Proyecto",
-    work_mode: "Trabajo",
-    zen_mode: "Zen",
-    title: "Título",
-    screensaver_mode: "Screensaver",
-    author: "Autor",
-    language: "Idioma",
-    theme: "Tema",
-
-    zen_mode_title: "Elimina todo lo que te distrae",
-    work_mode_title: "Mostrar solo citas seguras para el trabajo",
-    exit_zen_title: "Salir del modo Zen",
-
-    colors: "Colores",
-    base: "Base",
-    pink: "Rosa",
-    green: "Verde",
-    orange: "Naranja",
-    purple: "Morado",
-    blue: "Azul",
-    gray: "Gris",
-    color: "Colores aleatorios",
-
-    themes: "Temas",
-    retro: "Retro",
-    elegant: "Elegante",
-    festive: "Festivo",
-    bohemian: "Bohemio",
-    handwriting: "Manuscrita",
-    anaglyph: "Anaglifo",
-    whatsapp: "Whatsapp",
-    terminal: "Terminal",
-    frame: "Marco",
-    subtle: "Sutil",
-
-    system: "Sistema",
-    light: "Claro",
-    dark: "Oscuro",
-
-    "en-US": "Inglés (US)",
-    "es-ES": "Español (ES)",
-    "pt-BR": "Portugués (BR)",
-    "fr-FR": "Francés",
-    "it-IT": "Italiano",
-    random: "Lenguaje aleatorio",
-
-    themeFont: "Fuente del tema",
-  },
-  "pt-BR": {
-    document_title: "Relógio da Literatura",
-    add_quote: "Adicionar citação",
-    report_error: "Informar erro",
-    made_by: "Feito por",
-    project: "Projeto",
-    work_mode: "Trabalho",
-    zen_mode: "Zen",
-    screensaver_mode: "Screensaver",
-    title: "Título",
-    author: "Autor",
-    language: "Idioma",
-    theme: "Tema",
-
-    zen_mode_title: "Remova tudo o que te distrai",
-    work_mode_title: "Mostrar apenas citações seguras para o trabalho",
-    exit_zen_title: "Sair do modo Zen",
-
-    colors: "Cores",
-    base: "Base",
-    pink: "Rosa",
-    green: "Verde",
-    orange: "Laranja",
-    purple: "Roxo",
-    blue: "Azul",
-    gray: "Cinza",
-    retro: "Retrô",
-    color: "Cores aleatórias",
-
-    themes: "Temas",
-    elegant: "Elegante",
-    festive: "Festivo",
-    bohemian: "Boêmio",
-    handwriting: "Manuscrita",
-    anaglyph: "Anaglifo",
-    whatsapp: "Whatsapp",
-    terminal: "Terminal",
-    frame: "Moldura",
-    subtle: "Sutil",
-
-    system: "Sistema",
-    light: "Claro",
-    dark: "Escuro",
-
-    "en-US": "Inglês (US)",
-    "es-ES": "Espanhol (ES)",
-    "pt-BR": "Português (BR)",
-    "fr-FR": "Francês",
-    "it-IT": "Italiano",
-    random: "Idioma aleatório",
-
-    themeFont: "Fonte do tema",
-  },
-  "fr-FR": {
-    document_title: "Horloge de la Littérature",
-    add_quote: "Ajouter une citation",
-    report_error: "Informer l'erreur",
-    made_by: "Fait par",
-    project: "Projet",
-    work_mode: "Travail",
-    zen_mode: "Zen",
-    screensaver_mode: "Screensaver",
-    title: "Titre",
-    author: "Auteur",
-    language: "Langue",
-    theme: "Thème",
-
-    zen_mode_title: "Supprimez tout ce qui vous distrait",
-    work_mode_title:
-      "Afficher uniquement des citations appropriées pour le travail",
-    exit_zen_title: "Quitter le mode Zen",
-
-    colors: "Couleurs",
-    base: "Base",
-    pink: "Rose",
-    green: "Vert",
-    orange: "Orange",
-    purple: "Violet",
-    blue: "Bleu",
-    gray: "Gris",
-    color: "Couleurs aléatoires",
-
-    themes: "Thèmes",
-    retro: "Rétro",
-    elegant: "Élégant",
-    festive: "Festif",
-    bohemian: "Bohémien",
-    handwriting: "Manuscrit",
-    anaglyph: "Anaglyphe",
-    whatsapp: "Whatsapp",
-    terminal: "Terminal",
-    frame: "Bordure",
-    subtle: "Subtil",
-
-    system: "Système",
-    light: "Clair",
-    dark: "Sombre",
-
-    "en-US": "Anglais (US)",
-    "es-ES": "Espagnol (ES)",
-    "pt-BR": "Portugais (BR)",
-    "fr-FR": "Français",
-    "it-IT": "Italien",
-    random: "Langue aléatoire",
-
-    themeFont: "Police du thème",
-  },
-  "it-IT": {
-    document_title: "Orologio della Letteratura",
-    add_quote: "Aggiungi citazione",
-    report_error: "Informare l'errore",
-    made_by: "Realizzato da",
-    project: "Progetto",
-    work_mode: "Lavoro",
-    zen_mode: "Zen",
-    screensaver_mode: "Screensaver",
-    title: "Titolo",
-    author: "Autore",
-    addQuote: "Aggiungere una citazione",
-    language: "Lingua",
-    theme: "Tema",
-
-    zen_mode_title: "Rimuovi tutto ciò che ti distrae",
-    work_mode_title: "Mostra solo citazioni appropriate per il lavoro",
-    exit_zen_title: "Esci dalla modalità Zen",
-
-    colors: "Colori",
-    base: "Base",
-    pink: "Rosa",
-    green: "Verde",
-    orange: "Arancione",
-    purple: "Viola",
-    blue: "Blu",
-    gray: "Grigio",
-    color: "Colori casuali",
-
-    themes: "Temi",
-    retro: "Retro",
-    elegant: "Elegante",
-    festive: "Festivo",
-    bohemian: "Bohemien",
-    handwriting: "Manoscritto",
-    anaglyph: "Anaglifo",
-    whatsapp: "Whatsapp",
-    terminal: "Terminale",
-    frame: "Cornice",
-    subtle: "Sottile",
-
-    system: "Sistema",
-    light: "Chiaro",
-    dark: "Scuro",
-
-    "en-US": "Inglese (US)",
-    "es-ES": "Spagnolo (ES)",
-    "pt-BR": "Portoghese (BR)",
-    "fr-FR": "Francese",
-    "it-IT": "Italiano",
-    random: "Lingua casuale",
-
-    themeFont: "Font del tema ",
-  },
-} as const;
