+++ src/strings/fallbackQuotes.json
@@ -1,0 +1,112 @@
+{
+  "en-US": [
+    {
+      "quote_first": "Error ",
+      "quote_last": ": quote not found.",
+      "title": "Internet Explorer",
+      "author": "1995-2022",
+      "id": "",
+      "time": "",
+      "quote_time_case": "",
+      "sfw": "sfw"
+    },
+    {
+      "quote_first": "Captain's log:<br>We are still looking for a quote for ",
+      "quote_last": ".",
+      "title": "Moby Dick",
+      "author": "Captain Ahab",
+      "id": "",
+      "time": "",
+      "quote_time_case": "",
+      "sfw": "sfw"
+    }
+  ],
+  "es-ES": [
+    {
+      "quote_first": "Error ",
+      "quote_last": ": quote not found.",
+      "title": "Internet Explorer",
+      "author": "1995-2022",
+      "id": "",
+      "time": "",
+      "quote_time_case": "",
+      "sfw": "sfw"
+    },
+    {
+      "quote_first": "Bitácora del Capitán:<br>Seguimos buscando una cita para las ",
+      "quote_last": ".",
+      "title": "Moby Dick",
+      "author": "Captain Ahab",
+      "id": "",
+      "time": "",
+      "quote_time_case": "",
+      "sfw": "sfw"
+    }
+  ],
+  "pt-BR": [
+    {
+      "quote_first": "Erro ",
+      "quote_last": ": citação não encontrada.",
+      "title": "Internet Explorer",
+      "author": "1995-2022",
+      "id": "",
+      "time": "",
+      "quote_time_case": "",
+      "sfw": "sfw"
+    },
+    {
+      "quote_first": "Registro do capitão:<br>Ainda estamos procurando uma data para o ",
+      "quote_last": ".",
+      "title": "Moby Dick",
+      "author": "Capitão Ahab",
+      "id": "",
+      "time": "",
+      "quote_time_case": "",
+      "sfw": "sfw"
+    }
+  ],
+  "fr-FR": [
+    {
+      "quote_first": "Erreur ",
+      "quote_last": ": citation non trouvée.",
+      "title": "Internet Explorer",
+      "author": "1995-2022",
+      "id": "",
+      "time": "",
+      "quote_time_case": "",
+      "sfw": "sfw"
+    },
+    {
+      "quote_first": "Journal du capitaine:<br>Nous sommes toujours à la recherche d'una citation pour ",
+      "quote_last": ".",
+      "title": "Moby Dick",
+      "author": "Capitaine Achab",
+      "id": "",
+      "time": "",
+      "quote_time_case": "",
+      "sfw": "sfw"
+    }
+  ],
+  "it-IT": [
+    {
+      "quote_first": "Errore ",
+      "quote_last": ": citazione non trovata.",
+      "title": "Internet Explorer",
+      "author": "1995-2022",
+      "id": "",
+      "time": "",
+      "quote_time_case": "",
+      "sfw": "sfw"
+    },
+    {
+      "quote_first": "Diario del capitano:<br>Stiamo ancora cercando una data per il ",
+      "quote_last": ".",
+      "title": "Moby Dick",
+      "author": "Capitano Achab",
+      "id": "",
+      "time": "",
+      "quote_time_case": "",
+      "sfw": "sfw"
+    }
+  ]
+}
