+++ src/utils/utils.test.ts
@@ -1,0 +1,3 @@
+import { describe } from "vitest";
+
+describe("utils", () => {});
