+++ src/utils/utils.ts
@@ -1,92 +1,6 @@
-import { startScreensaver } from "./screensaver";
+import { startScreensaver } from "../modules/screensaver";
 import { isBooleanSettingTrue } from "./settings";
-import { Locale, Quote } from "./types";
-
-export const fallbackQuote = (quote: Partial<Quote>): Quote =>
-  ({
-    id: "",
-    time: "",
-    quote_time_case: "",
-    sfw: "sfw",
-    ...quote,
-  } as Quote);
-
-export const FALLBACK_QUOTES: Record<Locale, Quote[]> = {
-  "en-US": [
-    fallbackQuote({
-      quote_first: "Error ",
-      quote_last: ": quote not found.",
-      title: "Internet Explorer",
-      author: "1995-2022",
-    }),
-    fallbackQuote({
-      quote_first: "Captain's log:<br>We are still looking for a quote for ",
-      quote_last: ".",
-      title: "Moby Dick",
-      author: "Captain Ahab",
-    }),
-  ],
-  "es-ES": [
-    fallbackQuote({
-      quote_first: "Error ",
-      quote_last: ": quote not found.",
-      title: "Internet Explorer",
-      author: "1995-2022",
-    }),
-    fallbackQuote({
-      quote_first:
-        "Bitácora del Capitán:<br>Seguimos buscando una cita para las ",
-      quote_last: ".",
-      title: "Moby Dick",
-      author: "Captain Ahab",
-    }),
-  ],
-  "pt-BR": [
-    fallbackQuote({
-      quote_first: "Erro ",
-      quote_last: ": citação não encontrada.",
-      title: "Internet Explorer",
-      author: "1995-2022",
-    }),
-    fallbackQuote({
-      quote_first:
-        "Registro do capitão:<br>Ainda estamos procurando uma data para o ",
-      quote_last: ".",
-      title: "Moby Dick",
-      author: "Capitão Ahab",
-    }),
-  ],
-  "fr-FR": [
-    fallbackQuote({
-      quote_first: "Erreur ",
-      quote_last: ": citation non trouvée.",
-      title: "Internet Explorer",
-      author: "1995-2022",
-    }),
-    fallbackQuote({
-      quote_first:
-        "Journal du capitaine:<br>Nous sommes toujours à la recherche d'una citation pour ",
-      quote_last: ".",
-      title: "Moby Dick",
-      author: "Capitaine Achab",
-    }),
-  ],
-  "it-IT": [
-    fallbackQuote({
-      quote_first: "Errore ",
-      quote_last: ": citazione non trovata.",
-      title: "Internet Explorer",
-      author: "1995-2022",
-    }),
-    fallbackQuote({
-      quote_first:
-        "Diario del capitano:<br>Stiamo ancora cercando una data per il ",
-      quote_last: ".",
-      title: "Moby Dick",
-      author: "Capitano Achab",
-    }),
-  ],
-};
+import { Locale, Quote } from "../types";
 
 const INITIAL_THEME_FONT_SIZE: Record<string, number> = {
   handwriting: 90,
